Add files via upload
This is a bash script for automating the processing of samples through HISAT2, SAM tools, Stringtie, and output of data in a format fit for further ballgown analysis in R. It assumes that these command line softwares are installed and binaries are properly linked so that their commands can be invoked by script in any folder. It was written for analysing the JHU samples that can be found at ftp://ftp.ccb.jhu.edu/pub/RNAseq_protocol/. However, it can be modified to process other samples.